Move vote class code to a separate file
